kopi-debate-api/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
dependency-injector
pydantic
google-generativeai
